remove StaticContentSharedJvm.scala, scaladoc tweaks
